+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Change Log
 All notable changes to the "texty" extension will be documented in this file.
 
+## [0.0.8] - 2019-11-30
+### Added
+- Add `Open file in System default app` Command
+- Add `Open file in Browser` Command
+- Add `Open file in Browser...` Command
+- Add `Filter lines containing string` Command
+- Add `Filter lines containing string into new editor` Command
+- Add `Filter lines not containing string` Command
+- Add `Filter lines not containing string into new editor` Command
+- Add `Filter lines containing regex` Command
+- Add `Filter lines containing regex into new editor` Command
+- Add `Filter lines not containing regex` Command
+- Add `Filter lines not containing regex into new editor` Command
+- Add `Insert DateTime ISO Format` Command
+- Add `Insert Date ISO Format` Command
+- Add `Convert Date/Time to different format or timezone` Command
+- Add `Convert Date/Time to relative time` Command
+
 ## [0.0.7] - 2019-11-29
 ### Added
 - Add `Insert Short Id` Command